Don't automatically add charset to Accept headers, fixes #138
The charset parameter which was automatically added is invalid for some
content types whose specification demands a particular
charset (e.g. JSON, EDN), and thus this behavior renders cljs-ajax
broken when talking to some standards-compliant web servers (e.g. Yada).

NOTE: The behavior introduced in this commit is not backwards
compatible.  Of the four core formats, `:transit`, `:json`, `:text` and
`:raw`...

1. `:transit` and `:json` will now do the right thing and not send
   invalid charsets.
2. `:text` and `:raw` will continue to send `charset=utf-8`, as before.
3. If a `:content-type` is explicitly specified, it will be
   respected (unlike before, when it would have been changed and there
   was no way to prevent this).

Finally, the `Accept` headers sent will no longer include charsets.  Not
only does the validity of such parameters seem to be in question [1],
but even the seemingly-more-legitimate `Accept-Charset` seems to be
falling out of favor [2].

[1]: http://stackoverflow.com/a/7310784
[2]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Content_negotiation#The_Accept-Charset_header