Give better error messages in ajax-call with keywords as formats
ajax-call requires users to specify full request and response formats,
not just keywords. Because keywords are IFn's they fall into the ifn?
branch of the cond. This means that errors propagate further downstream
and are more confusing when they finally break.

This patch adds an assertion that keywords aren't passed as formats,
and throws a more informative error message if they are.

There are also some tweaks to the README to make the need to provide
full formats more clear.