Split out interceptors.cljc from core.cljc
One of the things I'm trying to solve at the moment is being able to have alternative json implementations.
This, in turn, would be helped by having json be in a different file from everything else. Whereas I've done
this before, in the case of EDN, that was at the cost of no longer being able to reference it within ajax.core.
In the case of EDN, it was being deprecated and forcing people to add a namespace was actually pretty attractive,
but there's no way I want to break everyone using JSON. The crucial function that needs extracting from core.cljc
to achieve this is map->ResponseFormat, but splitting out all the default interceptors seems like a reasonably
clean way of doing this.

This also helps with my long standing program of cutting core.cljc down. At the moment everything is in it,
which is a modularity nightmare. This forces a lot more separation and explicit linkage between the components.